Shorten a little

+++ curve25519-dalek/src/backend/serial/u64/scalar_verus.rs
@@ -94,7 +94,7 @@ verus! {
             if limbs.len() == 0 {
                 0
             } else {
-                limbs[0] + seq_to_nat(limbs.subrange(1 as int, limbs.len() as int)) * pow2(52)
+                limbs[0] + seq_to_nat(limbs.subrange(1, limbs.len() as int)) * pow2(52)
             }
         }
